+++ VadereSimulator/src/org/vadere/simulator/models/DynamicElementFactory.java
@@ -1,6 +1,7 @@
 package org.vadere.simulator.models;
 
 import org.jetbrains.annotations.NotNull;
+import org.vadere.state.attributes.Attributes;
 import org.vadere.state.scenario.DynamicElement;
 import org.vadere.state.scenario.Topography;
 import org.vadere.util.geometry.shapes.VPoint;
@@ -14,7 +15,16 @@ public interface DynamicElementFactory {
 	 * Note: Every attribute of the given element should be cloned for each individual in this
 	 * method, because some fields are individual.
 	 */
-	<T extends DynamicElement> DynamicElement createElement(VPoint position, int id, Class<T> type);
+	 <T extends DynamicElement> DynamicElement createElement(VPoint position, int id, Class<T> type);
+
+	/**
+	 * Override the DynamicElement Attributes given in the topography by the values given here.
+	 * Note: The Attribute values come from the source and are set by the SourceController
+	 * during DynamicElement Creation.
+	 */
+	<T extends DynamicElement> DynamicElement createElement(VPoint position, int id, Attributes attr, Class<T> type);
+
+
 
 	/**
 	 * Ensure all Ids are unique within the topography.