+++ VadereSimulator/src/org/vadere/simulator/models/reynolds/ReynoldsSteeringModel.java
@@ -112,10 +112,17 @@ public class ReynoldsSteeringModel implements MainModel {
 
 	@Override
 	public <T extends DynamicElement> Pedestrian createElement(VPoint position, int id, Class<T> type) {
+		return createElement(position, id, this.attributesPedestrian, type);
+	}
+
+	@Override
+	public <T extends DynamicElement> Pedestrian createElement(VPoint position, int id, Attributes attr, Class<T> type) {
+		AttributesAgent aAttr = (AttributesAgent)attr;
+
 		if (!Pedestrian.class.isAssignableFrom(type))
 			throw new IllegalArgumentException("RSM cannot initialize " + type.getCanonicalName());
 		AttributesAgent pedAttributes = new AttributesAgent(
-				attributesPedestrian, registerDynamicElementId(domain.getTopography(), id));
+				aAttr, registerDynamicElementId(domain.getTopography(), id));
 		Pedestrian result = create(position, pedAttributes);
 		return result;
 	}