+++ VadereSimulator/tests/org/vadere/simulator/control/TestSourceController.java
@@ -6,6 +6,7 @@ import org.vadere.simulator.control.factory.SingleSourceControllerFactory;
 import org.vadere.simulator.control.factory.SourceControllerFactory;
 import org.vadere.simulator.control.scenarioelements.SourceController;
 import org.vadere.simulator.models.DynamicElementFactory;
+import org.vadere.state.attributes.Attributes;
 import org.vadere.state.attributes.scenario.AttributesAgent;
 import org.vadere.state.attributes.scenario.AttributesSource;
 import org.vadere.state.attributes.scenario.SourceTestAttributesBuilder;
@@ -72,6 +73,15 @@ public class TestSourceController {
 
             @Override
             public <T extends DynamicElement> DynamicElement createElement(VPoint position, int id, Class<T> type) {
+                AttributesAgent att = new AttributesAgent(
+                        d.attributesPedestrian, registerDynamicElementId(null, id));
+                Pedestrian ped = new Pedestrian(att, d.random);
+                ped.setPosition(position);
+                return ped;
+            }
+
+            @Override
+            public <T extends DynamicElement> DynamicElement createElement(VPoint position, int id, Attributes attr, Class<T> type) {
                 AttributesAgent att = new AttributesAgent(
                         d.attributesPedestrian, registerDynamicElementId(null, id));
                 Pedestrian ped = new Pedestrian(att, d.random);